refactor: top emoji, share component 분리
